feat: add ActivitiesListView component
