+++ packages/apps/project/src/i18n/en.json
@@ -36,6 +36,8 @@
   "Project_Product": "Product",
   "Project_Duration": "Duration",
   "Project_ToInvoice": "To Invoice",
+  "Project_Reporting": "Reporting",
+  "Project_Activities": "Activities",
   "Project_SliceAction_SearchProject": "search projects",
   "Project_SliceAction_FetchProjectById": "fetch project by id",
   "Project_SliceAction_FetchProjectStatus": "fetch project status",
